NEW: start probabilistic transition
- I think the transition logic is okay... but we can't generate a deterministic policy from a probabilistic transition probability.. can we?
- Move transition prob to separate file since the logic is overly complex now...
